add cleanfrom CQRS

+++ Skeletons/CQRS/Clean CQRS/CleanFromScratch.API/CleanFromScratch.Appplication/Dishes/Commands/DeleteDish/DeleteDishesForRestaurantCommandHandler.cs
@@ -4,12 +4,14 @@ using CleanFromScratch.Domain.Constants;
 using CleanFromScratch.Domain.Entities;
 using CleanFromScratch.Domain.Exceptions;
 using CleanFromScratch.Domain.Repositories;
+using CleanFromScratch.Domain.Interfaces;
 
 namespace CleanFromScratch.Application.Dishes.Commands.DeleteDishes;
 
 public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService
     ) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
@@ -19,8 +21,8 @@ public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRest
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
         if (restaurant == null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
 
-        //if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
-        //    throw new ForbidException();
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbidException();
 
         await dishesRepository.Delete(restaurant.Dishes);
     }
